Add IPC Self-Repair Implant
This commit introduces a new self-repair implant for IPCs. The implant allows IPCs to slowly heal their brute and burn damage over time. The healing process is power inefficient, requiring IPCs to be near an APC to recharge. The implant is added to the list of medical designs that can be built with the Protolathe and Exosuit Fabricator.